modi sth

+++ mtools/np.py
@@ -40,7 +40,7 @@ def np_union_shuffle(a, b):
     return a[p], b[p]
 
 def np_repeat(data, nums):
-    return np.vstack(tuple([np.array(np.tile(row, (s, 1))) for row, s in zip(data, nums)]))
+    return np.vstack(tuple([np.array(np.tile(row, (s, 1)), dtype=data.dtype) for row, s in zip(data, nums)]))
 
 def rmse(a, b):
     return np.sqrt(np.mean((a-b)**2, axis=-1))
